Use canWait global to wait on Promise
The **canWait** global creates a callback function that waits within the
request lifecycle. We need to use this because can-connect brings in its
own Promise polyfill. So the first time on the server data is added to
a canWait request the Promise wrapper won't wrap can-connect's promises.
The second time it will (because it will be the global Promise at that
		point). So we can use a **canWait** callback that will tap into
the lifecycle.