Add can-connect/can/model to namespace
This makes it so that can-connect/can/model/ is added to the
can-util/namespace *only* if there isn't already a Model on the
namespace. can-model will **always** be added to the namespace. In this
way, can-connect/can/model/ acts as a fallback only when can-model isn't
already loaded.

This is for the can/legacy changes.