Adding a fallback to the steal.js script
When this test runs in canjs/canjs, the steal.js script is not located
in the same place. This adds a fallback to the script location. It is
also updated to use `steal.done()` and `System.import` since the code is no
longer in the steal.js script body, so it cannot use `require`.